DRILL-4047: review feedback. More doc; extracted classes
