+++ exec/java-exec/src/test/java/org/apache/drill/exec/store/dfs/TestFormatPluginOptionExtractor.java
@@ -1,3 +1,20 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.apache.drill.exec.store.dfs;
 
 import static org.junit.Assert.assertEquals;
@@ -8,23 +25,21 @@ import java.util.Collection;
 import org.apache.drill.common.config.DrillConfig;
 import org.apache.drill.common.scanner.RunTimeScan;
 import org.apache.drill.common.scanner.persistence.ScanResult;
-import org.apache.drill.exec.store.dfs.WorkspaceSchemaFactory.OptionExtractor;
-import org.apache.drill.exec.store.dfs.WorkspaceSchemaFactory.OptionsDescriptor;
 import org.apache.drill.exec.store.easy.text.TextFormatPlugin.TextFormatConfig;
 import org.junit.Test;
 
 import com.fasterxml.jackson.annotation.JsonTypeName;
 
 
-public class TestOptionExtractor {
+public class TestFormatPluginOptionExtractor {
 
   @Test
   public void test() {
     DrillConfig config = DrillConfig.create();
     ScanResult scanResult = RunTimeScan.fromPrescan(config);
-    OptionExtractor e = new OptionExtractor(scanResult);
-    Collection<OptionsDescriptor> options = e.getOptions();
-    for (OptionsDescriptor d : options) {
+    FormatPluginOptionExtractor e = new FormatPluginOptionExtractor(scanResult);
+    Collection<FormatPluginOptionsDescriptor> options = e.getOptions();
+    for (FormatPluginOptionsDescriptor d : options) {
       assertEquals(d.pluginConfigClass.getAnnotation(JsonTypeName.class).value(), d.typeName);
       switch (d.typeName) {
         case "text":
